+++ plugins/scaffolder-react/src/next/components/Stepper/createAsyncValidators.test.ts
@@ -406,4 +406,39 @@ describe('createAsyncValidators', () => {
 
     expect(validators.TagField).not.toHaveBeenCalled();
   });
+
+  it('should call validator for array object property from a custom field extension', async () => {
+    const schema: JsonObject = {
+      type: 'object',
+      properties: {
+        links: {
+          title: 'Links',
+          type: 'array',
+          items: {
+            type: 'object',
+            required: ['url', 'title', 'icon'],
+            properties: {
+              url: {
+                title: 'url',
+                description: 'url',
+                type: 'object',
+                'ui:field': 'CustomLinkField',
+              },
+            },
+          },
+        },
+      },
+    };
+    const validators = { CustomLinkField: jest.fn() };
+
+    const validate = createAsyncValidators(schema, validators, {
+      apiHolder: { get: jest.fn() },
+    });
+
+    await validate({
+      links: [{ url: 'http://my-url.spotify.com' }],
+    });
+
+    expect(validators.CustomLinkField).toHaveBeenCalled();
+  });
 });