+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # techdocs-cli-embedded-app
 
+## 0.2.80-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.5-next.2
+  - @backstage/plugin-techdocs-react@1.1.4-next.2
+  - @backstage/plugin-techdocs@1.6.0-next.2
+  - @backstage/core-app-api@1.6.0-next.2
+  - @backstage/plugin-catalog@1.9.0-next.2
+  - @backstage/core-plugin-api@1.5.0-next.2
+  - @backstage/app-defaults@1.2.1-next.2
+  - @backstage/cli@0.22.4-next.1
+  - @backstage/integration-react@1.1.11-next.2
+  - @backstage/test-utils@1.2.6-next.2
+  - @backstage/config@1.0.7-next.0
+
 ## 0.2.80-next.1
 
 ### Patch Changes