+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-bazaar
 
+## 0.2.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.5-next.2
+  - @backstage/plugin-catalog-react@1.4.0-next.2
+  - @backstage/plugin-catalog@1.9.0-next.2
+  - @backstage/core-plugin-api@1.5.0-next.2
+  - @backstage/cli@0.22.4-next.1
+
 ## 0.2.6-next.1
 
 ### Patch Changes