+++ plugins/catalog-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-catalog-node
 
+## 1.3.4-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.4.1-next.2
+
 ## 1.3.4-next.1
 
 ### Patch Changes