+++ plugins/catalog-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-react
 
+## 1.4.0-next.2
+
+### Patch Changes
+
+- 65454876fb2: Minor API report tweaks
+- Updated dependencies
+  - @backstage/core-components@0.12.5-next.2
+  - @backstage/core-plugin-api@1.5.0-next.2
+  - @backstage/plugin-permission-react@0.4.11-next.2
+  - @backstage/integration@1.4.3-next.0
+
 ## 1.4.0-next.1
 
 ### Patch Changes