+++ plugins/sonarqube-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-sonarqube-react
 
+## 0.1.4-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.0-next.2
+
 ## 0.1.4-next.1
 
 ### Patch Changes