+++ plugins/sonarqube/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-sonarqube
 
+## 0.6.5-next.2
+
+### Patch Changes
+
+- 65454876fb2: Minor API report tweaks
+- Updated dependencies
+  - @backstage/core-components@0.12.5-next.2
+  - @backstage/plugin-catalog-react@1.4.0-next.2
+  - @backstage/core-plugin-api@1.5.0-next.2
+  - @backstage/plugin-sonarqube-react@0.1.4-next.2
+
 ## 0.6.5-next.1
 
 ### Patch Changes